Implement companion object constant access (#247)
Intercept Int/Long/Char companion constants (MIN_VALUE, MAX_VALUE) in
visitPropertyAccessExpression before the qualifier FirResolvedQualifier
is visited, emitting them as integer/char literals. Also widens IntLit
from Int to Long to support Long.MIN_VALUE and Long.MAX_VALUE.

+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/conversion/StdLibConverter.kt
@@ -22,7 +22,7 @@ private fun VariableEmbedding.sameSize(): ExpEmbedding =
 private fun VariableEmbedding.increasedSize(amount: Int): ExpEmbedding =
     EqCmp(
         FieldAccess(this, CollectionSizeFieldEmbedding),
-        OperatorExpEmbeddings.AddIntInt(Old(FieldAccess(this, CollectionSizeFieldEmbedding)), IntLit(amount)),
+        OperatorExpEmbeddings.AddIntInt(Old(FieldAccess(this, CollectionSizeFieldEmbedding)), IntLit(amount.toLong())),
     )
 
 sealed interface StdLibReceiverInterface {
